Feature/readme

+++ js/panel-admin/p-admin-usuarios.js
@@ -205,10 +205,44 @@ function editarUsuario(e) {
 }
 
 function eliminarUsuario(idUsuario) {
+  // Verificar si el ID es cero
+  if (idUsuario === 0) {
+    Swal.fire({
+      icon: 'error',
+      title: 'Oops...',
+      text: 'No se puede eliminar el usuario administrador.',
+    });
+    return; // No hace nada si el ID es cero
+  }
+
+  let usuarioAEliminar = usuariosRegistrados.find(function (usuario) {
+    return usuario.id === idUsuario;
+  });
+
+  if (!usuarioAEliminar) {
+    alert('El usuario no existe.');
+    return; // No hace nada si el usuario no existe
+  }
+
+  let confirmacion = confirm(`¿Esta seguro que desea eliminar al usuario ${usuarioAEliminar.nombre}?`);
+
+  if (!confirmacion) {
+    return; // No hace nada si el usuario cancela la eliminación
+  }
+
+  // Filtrar el arreglo de usuarios para eliminar el usuario con el ID correspondiente
   usuariosRegistrados = usuariosRegistrados.filter(function (usuario) {
+    Swal.fire({
+      icon: 'success',
+      title: `${usuarioAEliminar.nombre} eliminado/a`,
+    });
     return usuario.id !== idUsuario;
   });
+
+  // Actualizar el almacenamiento local con el nuevo arreglo de usuarios
   localStorage.setItem('usuarios', JSON.stringify(usuariosRegistrados));
+
+  // Cargar los usuarios nuevamente (asumiendo que cargaUsuarios está definido en otro lugar)
   cargaUsuarios(0);
 }
